use PreviewValue and handle complex enum values
